Split up MKS_RUMBA32 into two variants (#16781)
