+++ Marlin/src/core/boards.h
@@ -310,19 +310,20 @@
 #define BOARD_BEAST                   4200  // STM32F4xxVxT6 Libmaple-based STM32F4 controller
 #define BOARD_GENERIC_STM32F4         4201  // STM32 STM32GENERIC-based STM32F4 controller
 #define BOARD_ARMED                   4202  // Arm'ed STM32F4-based controller
-#define BOARD_RUMBA32                 4203  // RUMBA32 STM32F4-based controller
-#define BOARD_BLACK_STM32F407VE       4204  // BLACK_STM32F407VE
-#define BOARD_BLACK_STM32F407ZE       4205  // BLACK_STM32F407ZE
-#define BOARD_STEVAL_3DP001V1         4206  // STEVAL-3DP001V1 3D PRINTER BOARD
-#define BOARD_BTT_SKR_PRO_V1_1        4207  // BigTreeTech SKR Pro v1.1 (STM32F407ZG)
-#define BOARD_BTT_BTT002_V1_0         4208  // BigTreeTech BTT002 v1.0 (STM32F407VE)
-#define BOARD_BTT_GTR_V1_0            4209  // BigTreeTech GTR v1.0 (STM32F407IGT)
-#define BOARD_LERDGE_K                4210  // Lerdge K (STM32F407ZG)
-#define BOARD_LERDGE_X                4211  // Lerdge X (STM32F407VE)
-#define BOARD_VAKE403D                4212  // VAkE 403D (STM32F446VET6)
-#define BOARD_FYSETC_S6               4213  // FYSETC S6 board
-#define BOARD_FLYF407ZG               4214  // FLYF407ZG board (STM32F407ZG)
-#define BOARD_MKS_ROBIN2              4215  // MKS_ROBIN2 (STM32F407ZE)
+#define BOARD_RUMBA32_AUS3D           4203  // RUMBA32 STM32F446VET6 based controller from Aus3D
+#define BOARD_RUMBA32_MKS             4204  // RUMBA32 STM32F446VET6 based controller from Makerbase
+#define BOARD_BLACK_STM32F407VE       4205  // BLACK_STM32F407VE
+#define BOARD_BLACK_STM32F407ZE       4206  // BLACK_STM32F407ZE
+#define BOARD_STEVAL_3DP001V1         4207  // STEVAL-3DP001V1 3D PRINTER BOARD
+#define BOARD_BTT_SKR_PRO_V1_1        4208  // BigTreeTech SKR Pro v1.1 (STM32F407ZG)
+#define BOARD_BTT_BTT002_V1_0         4209  // BigTreeTech BTT002 v1.0 (STM32F407VE)
+#define BOARD_BTT_GTR_V1_0            4210  // BigTreeTech GTR v1.0 (STM32F407IGT)
+#define BOARD_LERDGE_K                4211  // Lerdge K (STM32F407ZG)
+#define BOARD_LERDGE_X                4212  // Lerdge X (STM32F407VE)
+#define BOARD_VAKE403D                4213  // VAkE 403D (STM32F446VET6)
+#define BOARD_FYSETC_S6               4214  // FYSETC S6 board
+#define BOARD_FLYF407ZG               4215  // FLYF407ZG board (STM32F407ZG)
+#define BOARD_MKS_ROBIN2              4216  // MKS_ROBIN2 (STM32F407ZE)
 
 //
 // ARM Cortex M7
