+++ Marlin/src/pins/lpc1768/pins_BTT_SKR_V1_1.h
@@ -230,4 +230,4 @@
 #endif
 
 // Include common SKR pins
-#include "pins_BTT_SKR.h"
+#include "pins_BTT_SKR_common.h"
