+++ Marlin/src/pins/lpc1768/pins_BTT_SKR_V1_3.h
@@ -369,4 +369,4 @@
  */
 
 // Include common SKR pins
-#include "pins_BTT_SKR.h"
+#include "pins_BTT_SKR_common.h"
