+++ Marlin/src/pins/lpc1768/pins_BTT_SKR_V1_4.h
@@ -153,7 +153,7 @@
 //
 // Include common SKR pins
 //
-#include "pins_BTT_SKR.h"
+#include "pins_BTT_SKR_common.h"
 
 //
 // Software SPI pins for TMC2130 stepper drivers
