+++ Marlin/src/pins/stm32/pins_RUMBA32_common.h
@@ -21,17 +21,23 @@
  */
 #pragma once
 
+/**
+ * Common pin assignments for all RUMBA32 boards
+ */
+
 #ifndef STM32F4
   #error "Oops! Select an STM32F4 board in 'Tools > Board.'"
 #elif HOTENDS > 3 || E_STEPPERS > 3
-  #error "RUMBA32 supports up to 3 hotends / E-steppers."
+  #error "RUMBA32 boards support up to 3 hotends / E-steppers."
 #endif
 
 #define RUMBA32_V1_0
-#define BOARD_INFO_NAME      "RUMBA32"
 #define DEFAULT_MACHINE_NAME BOARD_INFO_NAME
 
 //#define I2C_EEPROM
+#ifdef E2END
+  #undef E2END
+#endif
 #define E2END 0xFFF // 4KB
 
 //
@@ -119,12 +125,25 @@
 //
 // LCD / Controller
 //
-#define LCD_PINS_RS        PE10
-#define LCD_PINS_ENABLE    PE9
-#define LCD_PINS_D4        PE12
-#define LCD_PINS_D5        PE13
-#define LCD_PINS_D6        PE14
-#define LCD_PINS_D7        PE15
-#define BTN_EN1            PB1
-#define BTN_EN2            PB2
-#define BTN_ENC            PE7
+#if HAS_SPI_LCD
+
+  #define BTN_EN1          PB2
+  #define BTN_EN2          PB1
+  #define BTN_ENC          PE7
+
+  #define LCD_PINS_RS      PE10
+  #define LCD_PINS_ENABLE  PE9
+  #define LCD_PINS_D4      PE12
+
+  #if ENABLED(MKS_MINI_12864)
+    #define DOGLCD_CS      PE13
+    #define DOGLCD_A0      PE14
+  #endif
+
+  #if ENABLED(ULTIPANEL)
+    #define LCD_PINS_D5    PE13
+    #define LCD_PINS_D6    PE14
+    #define LCD_PINS_D7    PE15
+  #endif
+
+#endif
