+++ platformio.ini
@@ -822,9 +822,9 @@ monitor_speed = 500000
 upload_protocol = dfu
 
 #
-# MKS RUMBA32(add TMC2208/2209 UART interface and AUX-1)
+# MKS RUMBA32 (adds TMC2208/2209 UART interface and AUX-1)
 #
-[env:mks_rumba32]
+[env:rumba32_mks]
 platform      = ststm32
 board         = rumba32_f446ve
 build_flags   = ${common.build_flags}
